Implemented the allocate function for the Page Level Allocator and replaced UAddr with usize since the latter is always a pointer sized unsigned integer which is what UAddr was meant to be anyway.

+++ charlotte_core/src/arch/x86_64/memory/page_map/page_table/mod.rs
@@ -23,8 +23,8 @@ pub enum PageTableLevel {
 }
 
 const N_PT_ENTRIES: usize = 512;
-const LARGE_PAGE_NFRAMES: u64 = 512;
-const HUGE_PAGE_NFRAMES: u64 = 512 * 512;
+const LARGE_PAGE_NFRAMES: usize = 512;
+const HUGE_PAGE_NFRAMES: usize = 512 * 512;
 
 #[repr(C, align(4096))]
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
